requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
APScheduler==3.10.4
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
cryptography==41.0.4
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
Faker==19.6.2
fastapi==0.103.2
greenlet==3.0.0
h11==0.14.0
httpcore==0.18.0
httptools==0.6.0
httpx==0.25.0
idna==3.4
iniconfig==2.0.0
install==1.3.5
Mako==1.2.4
MarkupSafe==2.1.3
outline-vpn-api==6.2.0
packaging==23.2
passlib==1.7.4
Pillow==10.0.1
pipdeptree==2.13.0
pluggy==1.3.0
postgres==4.0
psycopg2-binary==2.9.9
psycopg2-pool==1.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyotp==2.9.0
pytest==7.4.2
pytest-asyncio==0.21.1
pytest-is-running==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.21
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.0.6
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3
click